refactor : 앤드포인트 기준 파일 변경 #16
