+++ src/server.ts
@@ -4,7 +4,6 @@ import cors from 'cors'
 import lecturesRouter from './routers/lectures/lecturesRouter.js'
 import healthRouter from './routers/health/healthRouter.js'
 import recruitRouter from './routers/recruit/recruitRouter.js'
-import recruitManageRouter from './routers/recruit/manage/recruitManageRouter.js'
 
 const app = express()
 app.use(express.json())
@@ -26,7 +25,6 @@ app.use(
 // NOTE:  넘어간 이후엔 `/lectures`를 제외한 엔드포인트만 쓰시면 됩니다
 app.use('/lectures', lecturesRouter)
 app.use('/recruitments', recruitRouter)
-app.use('/recruitManage', recruitManageRouter)
 app.use('/', healthRouter) // NOTE: 이 라우터는 가장 마지막에 있어야 합니다
 
 app.listen(3000, () => console.log('----server is on 3000'))
